requirements.txt
pymupdf
openai==0.28
python-dotenv
